Float return value from magnitude

+++ Pinta.Core/Algorithms/Utility.cs
@@ -29,7 +29,7 @@ public static class Utility
 	public static double MagnitudeSquared (PointD p)
 		=> Mathematics.MagnitudeSquared (p.X, p.Y);
 
-	public static double Magnitude (this PointF point)
+	public static float Magnitude (this PointF point)
 		=> Mathematics.Magnitude (point.X, point.Y);
 
 	public static double Magnitude (this PointD point)
